\name{mapping_data}
\alias{mapping_data}
\title{Convert Data to Other Corresponding Value}
\description{
}
\usage{
mapping_data(data=NULL, col_name=NULL, ori_tag=NULL, mapping_tag=NULL, other_tag=NULL)
}
\arguments{
\item{data}{data.frame: your input genomic data.}
\item{col_name}{character: column name.}
\item{ori_tag}{vector: a list of values contained in your input genomic data.}
\item{mapping_tag}{vector: a list of replaced values.}
\item{other_tag}{any basic data type or NULL (default is NULL): values contained in ori_tag but not in your input data.}
}
\examples{
}
