\name{MOSplot}
\alias{MOSplot}
\title{A Package For Creating MOSplot (mosquito coil plot) / Spiral Graph}
\description{
MOSplot accepts data table as input which contains at least three columns: panel id, fragment id and the corresponding value. Then, MOSplot creates spiral graph in these four steps: Firstly, graph parameters including graph layout setting, panel setting, grid setting, tag setting and geometric element setting are loaded and validated; Next, a blank canvas, background and base line are rendered in turn; After that, users can map their data to canvas with different geometric element including dot, line, bar, arrow-bar and trapezoid, or even user-defined element by x-axis and y-axis coordinate transformation; Lastly, shade region can be added to the top layer on the graph if the users want to highlight some regions.
}
\author{
\strong{Contact}: \email{shiwy@im.ac.cn}
}
\seealso{
Main page: \url{https://github.com/friendevangelion/MOSplot}
}